# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DamascusTransit API",
    "description": "Real-time transit tracking and fleet management — multi-tenant SaaS edition.\n\n## Authentication\n\nMost read endpoints are public but require an `?operator=<slug>` query parameter.\n1. **POST /api/v1/auth/login** — exchange email/password for a JWT token\n2. Include the token as `Authorization: Bearer <token>` on protected endpoints\n\nRoles: `super_admin` · `admin` · `dispatcher` · `driver` · `viewer`\n\nTokens expire after 24 hours.",
    "contact": {
      "name": "Damascus Transit System",
      "email": "support@damascustransit.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/routes/{route_id}": {
      "get": {
        "tags": [
          "routes"
        ],
        "summary": "Get Route",
        "description": "Get single route details.",
        "operationId": "get_route_api_routes__route_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Id"
            }
          },
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/routes/{route_pk}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Update Route",
        "description": "Update a route's business fields (fare, names, timing, active flag).",
        "operationId": "admin_update_route_api_admin_routes__route_pk__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_pk",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Route Pk"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pay/qr/{vehicle_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Get Vehicle Qr",
        "description": "Return the signed QR payload for a vehicle (driver/operator/admin).\n\nThe driver app renders this as a QR code; operators may also print it\nas the permanent in-vehicle sticker. Only approved vehicles get a code.",
        "operationId": "get_vehicle_qr_api_pay_qr__vehicle_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pay/initiate": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Initiate Payment",
        "description": "Passenger scanned a vehicle QR — create a pending fare payment.\n\nAnti-tamper rules:\n* QR signature must verify (counterfeit stickers rejected).\n* When the vehicle's assigned route has a fixed fare (bus/microbus),\n  the paid amount MUST equal the route fare — the client cannot\n  manipulate the price. Taxis (no fixed fare) pass the metered amount.",
        "operationId": "initiate_payment_api_pay_initiate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInitiateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentInitiateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pay/status/{payment_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Payment Status",
        "description": "Poll a payment's state (the UUID itself is the access capability).",
        "operationId": "payment_status_api_pay_status__payment_id__get",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pay/webhook/shamcash": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Shamcash Webhook",
        "description": "Sham Cash server-to-server confirmation callback (live mode).\n\nHMAC-SHA256 over the raw body with SHAM_CASH_WEBHOOK_SECRET,\nconstant-time compared. Fails closed when the secret is unset.",
        "operationId": "shamcash_webhook_api_pay_webhook_shamcash_post",
        "parameters": [
          {
            "name": "x-shamcash-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Shamcash-Signature"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShamCashWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pay/sandbox/confirm": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Sandbox Confirm",
        "description": "SANDBOX ONLY — simulate the Sham Cash confirmation callback.\n\nDisabled entirely in live mode. Lets the team demo the full passenger\npayment loop before merchant credentials exist.",
        "operationId": "sandbox_confirm_api_pay_sandbox_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShamCashWebhookPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/payments": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "List Payments",
        "description": "Operator-scoped payment ledger for the admin dashboard.",
        "operationId": "list_payments_api_admin_payments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentInitiateRequest": {
        "properties": {
          "qr": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 16,
            "title": "Qr"
          },
          "amount_syp": {
            "type": "integer",
            "maximum": 1000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount Syp"
          }
        },
        "type": "object",
        "required": [
          "qr",
          "amount_syp"
        ],
        "title": "PaymentInitiateRequest",
        "description": "Passenger scanned a vehicle QR and wants to pay the fare."
      },
      "PaymentInitiateResponse": {
        "properties": {
          "payment_id": {
            "type": "string",
            "title": "Payment Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "amount_syp": {
            "type": "integer",
            "title": "Amount Syp"
          },
          "vehicle_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Code"
          },
          "deeplink": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deeplink"
          },
          "sandbox": {
            "type": "boolean",
            "title": "Sandbox",
            "default": true
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "status",
          "amount_syp"
        ],
        "title": "PaymentInitiateResponse"
      },
      "PaymentStatusResponse": {
        "properties": {
          "payment_id": {
            "type": "string",
            "title": "Payment Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "amount_syp": {
            "type": "integer",
            "title": "Amount Syp"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "sandbox": {
            "type": "boolean",
            "title": "Sandbox",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "status",
          "amount_syp"
        ],
        "title": "PaymentStatusResponse"
      },
...
      "RouteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "route_id": {
            "type": "string",
            "title": "Route Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "route_type": {
            "type": "string",
            "title": "Route Type"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "avg_duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration Min"
          },
          "fare_syp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fare Syp"
          },
          "stop_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "route_id",
          "name",
          "name_ar",
          "route_type"
        ],
        "title": "RouteResponse"
      },
      "RouteUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "name_ar": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Ar"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^#[0-9a-fA-F]{6}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "avg_duration_min": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1440.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration Min"
          },
          "fare_syp": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fare Syp"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "RouteUpdate"
      },
...
      "ShamCashWebhookPayload": {
        "properties": {
          "payment_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Payment Id"
          },
          "provider_ref": {
            "type": "string",
            "maxLength": 128,
            "minLength": 4,
            "title": "Provider Ref"
          },
          "result": {
            "type": "string",
            "enum": [
              "success",
              "failure"
            ],
            "title": "Result"
          },
          "amount_syp": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Syp"
          },
          "payer_hint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Payer Hint"
          }
        },
        "type": "object",
        "required": [
          "payment_id",
          "provider_ref",
          "result",
          "amount_syp"
        ],
        "title": "ShamCashWebhookPayload",
        "description": "Callback body from Sham Cash (or the sandbox simulator)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}